separando a lógica de gerar as cartas da tela
