+++ memorygame/lib/components/cardList.dart
@@ -1,10 +1,9 @@
 import 'catcard.dart';
 import 'package:flutter/material.dart';
-import 'dart:math';
-
+import 'package:memorygame/models/memorygame.dart';
 class CardList extends StatelessWidget {
   final int numCards;
-  final List<int> imagePaths = [100, 102, 200, 201, 204, 214, 400, 404];
+  final MemoryGameManager _gameManager = MemoryGameManager();
 
   CardList({required this.numCards}) {
     if (numCards % 2 != 0) {
@@ -12,21 +11,9 @@ class CardList extends StatelessWidget {
     }
   }
 
-  List<int> generateImageList(int numCards) {
-    List<int> images = [];
-    Random rand = Random();
-    while (images.length < numCards) {
-      int index = rand.nextInt(imagePaths.length);
-      images.add(imagePaths[index]);
-      images.add(imagePaths[index]); // Add the pair
-    }
-    images.shuffle();
-    return images;
-  }
-
   @override
   Widget build(BuildContext context) {
-    List<int> images = generateImageList(numCards);
+    List<int> images = _gameManager.generateImageList(numCards);
 
     return GridView.builder(
       gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(
@@ -39,6 +26,7 @@ class CardList extends StatelessWidget {
       itemBuilder: (context, index) {
         return CatCard(
           statusCode: images[index],
+          index: index,
         );
       },
     );
